pyproject.toml
[tool.poetry]
name = "ollama-langchain"
version = "0.1.0"
description = ""
authors = ["Manuelji <manueljl22@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.0.353"
streamlit = "^1.29.0"
python-dotenv = "^1.0.0"
pypdf2 = "^3.0.1"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
